+++ packages/better_player/lib/src/configuration/player_log_level.dart
@@ -1,0 +1,16 @@
+enum PlayerLogLevel {
+  /// 0 — most verbose
+  debug,
+
+  /// 1
+  info,
+
+  /// 2
+  warning,
+
+  /// 3
+  error,
+
+  /// 4 — disables all logs
+  none,
+}